Add a helper to connect to Xpra inside the Docker container
Images generated by 'docker commit' actually do not inherit the CMD.
Therefore, when we try to commit the public key into the image, we will
lose the default command.

We could now play nasty little sed games with the output of 'docker
inspect', but the better way is to use a shared directory (useful for
data exchange in any case) for the container to pick up the public key:
that way, the image will actually not contain any default public key
(and can therefore not be forgotten to be replaced).

We will keep the change here, however, for the historical record.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>