Put the Xpra server into the background; open a root shell in foreground
Earlier, we tried to make things more convenient by starting the Xpra
server in the foreground so that quitting the last window in Xpra would
quit the container.

This was misguided, however, because Xpra does *not* quit when the last
window quits.

There was another problem we needed to solve somehow: due to our
reluctance to generate hard-to-remember passwords, we no longer could sudo
easily.

Both problems are addressed by opening an interactive root shell instead.
That way, we can -- if we want to -- add more settings via the console,
such as setting a password, or installing packages directly as root.

It is also a boon when debugging.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>