Switch from an ssh-based to a unix sockets-based method
Xpra has several modes how a client can connect to a server. The most
common is ssh because it provides very decent encryption without much
hassle between a client machine and a host machine.

However, in docker-desktop's scenario, the host machine is actually
running on the very same machine. So unix sockets are a much better
alternative because they provide security based on the file system,
and they are easier to set up (apart from having to start the Docker
container via sudo, this method does not require sudo because we
do not have to figure out which port the sshd is redirected to).

This also opens the door for "throw-away" containers, i.e. containers
whose sole purpose is to live as long as a given application is running
and after that be safely forgotten, with all its intermediate files.

Also, by making it very easy to have multiple containers, all running
an Xpra server inside, it is much easier to insulate applications,
or even different sessions of the same application, from one another.
Think e.g. surfing to a website whose advertisers read your Facebook
cookies.

These unix sockets can be accessed via ssh from a remote machine, too:

	REMOTE_SOCAT='socat - UNIX:docker-desktop/shared/docker-desktop-10'
	socat UNIX-LISTEN:$(hostname)-10 \
		EXEC:"ssh host-machine '$REMOTE_SOCAT'" &
	xpra --socket-dir=$(pwd) attach :10

In plain English: this uses a local and a remote socat, connected via ssh,
to allow a local Xpra to connect to the remote Docker container's Xpra.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>